把不同的函数放到不同的目录下。区分qstat和qmgr的功能
