--- connection/connection.go
@@ -1,1 +1,0 @@
-package connection
